Added type on Pin definitions

+++ pins_app.h
@@ -5,13 +5,13 @@
 #define CE_PIN 7
 #define CSN_PIN 8
 
-#define PHOTOCELL_DATA_PIN      2       //Interrupt attachable
-#define ANALOG_ENABLE_PIN       3       //Interrupt attachable
-#define OUTPUT_SIGNAL_PIN       4
-#define TM1637_CLK_PIN          6
+#define PHOTOCELL_DATA_PIN      (uint8_t) 2       //Interrupt attachable
+#define ANALOG_ENABLE_PIN       (uint8_t) 3       //Interrupt attachable
+#define OUTPUT_SIGNAL_PIN       (uint8_t) 4
+#define TM1637_CLK_PIN          (uint8_t) 6
 #define TM1637_DIO_PIN          A3
-#define LED1_PIN                9
-#define LED2_PIN                10
+#define LED1_PIN                (uint8_t) 9
+#define LED2_PIN                (uint8_t) 10
 #define LIPO_LEVEL_PIN          A0
